turn template into a library; turn its compositions into projects
- libraries/REBUS/REBUS.h is the old compositions/template/render.cpp
  with the JACK and SNDFILE backends removed (which only worked for
  trivial examples; if you want you can get them from earlier git commits)

- examples/REBUS/composition-api/ is the old compositions/template/example.h
  turned into an example project, using the new library

- projects/* are the other compostions using the template, now individual
  and using the new library; sketch.js and settings.json are mostly
  boilerplate repetitions

- how to deploy:

  1. make a backup of the sd card just in case
  2. `make -C libraries install`
  3. `make -C examples install`
  4. `ssh root@bela.local nano /root/Bela/examples/order.json`
     and add REBUS to the list
  5. reboot the board to reload the libraries in the IDE
  6. `make -C projects zip`
  7. upload desired projects to the IDE via web interface
