--- compositions/Makefile
@@ -1,5 +1,0 @@
-all: template.zip
-
-%.zip: %/ %/* Makefile
-	rm -rf $@ $</build/
-	zip -9 -r $@ $<
