+++ examples/Makefile
@@ -1,0 +1,6 @@
+default:
+	@echo "run make install to copy to bela.local"
+
+install:
+	ssh root@bela.local rm -rf /root/Bela/examples/REBUS
+	scp -r REBUS root@bela.local:/root/Bela/examples/
