+++ projects/Makefile
@@ -1,0 +1,8 @@
+DIRS = $(shell ls -d */)
+ZIPS = $(patsubst %/,%.zip,$(DIRS))
+
+zip: $(ZIPS)
+
+%.zip: %/ %/* Makefile
+	rm -rf $@ $</build/
+	zip -9 -r $@ $<
