+++ projects/i-spectral/settings.json
@@ -1,0 +1,34 @@
+{
+  "fileName": "render.cpp",
+  "CLArgs": {
+    "-p": "64",
+    "-C": "8",
+    "-B": "16",
+    "-N": 1,
+    "-G": 0,
+    "-M": "0",
+    "-H0,": "-6",
+    "-H1,": "-6",
+    "-H2,": "-6",
+    "-H3,": "-6",
+    "-H4,": "-6",
+    "-H5,": "-6",
+    "-H6,": "-6",
+    "-H7,": "-6",
+    "-I0,": "10",
+    "-I1,": "10",
+    "-I2,": "10",
+    "-I3,": "10",
+    "-I4,": "10",
+    "-I5,": "10",
+    "-I6,": "10",
+    "-I7,": "10",
+    "user": "",
+    "make": "",
+    "-X": "0",
+    "audioExpander": "0",
+    "-Y": "",
+    "-Z": "",
+    "--disable-led": "0"
+  }
+}
