+++ projects/spectrifier/square_table.h
